wip
- Change Gradle setup to be more appropriate for a library instead of an application
- Upgrade Kotlin and ktlint versions
- Change namespace to no.novari.*
- Replaced RestTemplate with RestClient - modern synchronous web client

** OAuth intercepter must be tested **
