--- src/main/kotlin/no/fintlabs/gateway/instance/exception/AbstractInstanceRejectedException.kt
@@ -1,5 +1,0 @@
-package no.fintlabs.gateway.instance.exception
-
-abstract class AbstractInstanceRejectedException(
-    message: String,
-) : RuntimeException(message)
