--- src/main/kotlin/no/fintlabs/gateway/instance/exception/FileUploadException.kt
@@ -1,9 +1,0 @@
-package no.fintlabs.gateway.instance.exception
-
-import no.fintlabs.gateway.instance.model.File
-
-class FileUploadException(
-    val file: File,
-    postResponse: String,
-    cause: Throwable? = null,
-) : RuntimeException("Could not post file=$file. POST response='$postResponse'", cause)
