--- src/main/kotlin/no/fintlabs/gateway/instance/exception/NoIntegrationException.kt
@@ -1,9 +1,0 @@
-package no.fintlabs.gateway.instance.exception
-
-import no.fintlabs.gateway.instance.model.SourceApplicationIdAndSourceApplicationIntegrationId
-
-class NoIntegrationException(
-    val sourceApplicationIdAndSourceApplicationIntegrationId: SourceApplicationIdAndSourceApplicationIntegrationId,
-) : RuntimeException(
-        "Count not find integration for ${sourceApplicationIdAndSourceApplicationIntegrationId.sourceApplicationId}",
-    )
