+++ src/main/kotlin/no/novari/gateway/instance/exception/AbstractInstanceRejectedException.kt
@@ -1,0 +1,5 @@
+package no.novari.gateway.instance.exception
+
+abstract class AbstractInstanceRejectedException(
+    message: String,
+) : RuntimeException(message)
