+++ src/main/kotlin/no/novari/gateway/instance/exception/IntegrationDeactivatedException.kt
@@ -1,0 +1,7 @@
+package no.novari.gateway.instance.exception
+
+import no.novari.gateway.instance.model.Integration
+
+class IntegrationDeactivatedException(
+    val integration: Integration,
+) : RuntimeException("Integration is deactivated: $integration")
