+++ src/main/kotlin/no/novari/gateway/instance/exception/NoIntegrationException.kt
@@ -1,0 +1,9 @@
+package no.novari.gateway.instance.exception
+
+import no.novari.gateway.instance.model.SourceApplicationIdAndSourceApplicationIntegrationId
+
+class NoIntegrationException(
+    val sourceApplicationIdAndSourceApplicationIntegrationId: SourceApplicationIdAndSourceApplicationIntegrationId,
+) : RuntimeException(
+        "Count not find integration for ${sourceApplicationIdAndSourceApplicationIntegrationId.sourceApplicationId}",
+    )
