+++ src/main/kotlin/no/novari/gateway/instance/model/ArchiveCaseIdRequestParams.kt
@@ -1,0 +1,6 @@
+package no.novari.gateway.instance.model
+
+data class ArchiveCaseIdRequestParams(
+    val sourceApplicationId: Long,
+    val sourceApplicationInstanceId: String,
+)
