+++ src/main/kotlin/no/novari/gateway/instance/model/SourceApplicationIdAndSourceApplicationIntegrationId.kt
@@ -1,0 +1,6 @@
+package no.novari.gateway.instance.model
+
+data class SourceApplicationIdAndSourceApplicationIntegrationId(
+    val sourceApplicationId: Long,
+    val sourceApplicationIntegrationId: String,
+)
