+++ src/main/kotlin/no/novari/gateway/instance/model/instance/InstanceObject.kt
@@ -1,0 +1,6 @@
+package no.novari.gateway.instance.model.instance
+
+data class InstanceObject(
+    val valuePerKey: Map<String, String> = HashMap(),
+    val objectCollectionPerKey: MutableMap<String, Collection<InstanceObject>> = HashMap(),
+)
