+++ src/main/kotlin/no/novari/gateway/instance/validation/InstanceValidationException.kt
@@ -1,0 +1,5 @@
+package no.novari.gateway.instance.validation
+
+class InstanceValidationException(
+    val validationErrors: List<InstanceValidationService.Error>,
+) : RuntimeException("Instance validation errors: $validationErrors")
